requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.15
helium>=3.2.3
selenium>=4.10.0
python-dotenv==0.19.2
requests>=2.25.1
pydantic>=2.0.0
beautifulsoup4>=4.12.2
aiohttp>=3.8.4
brotli